XMLHttpRequest/tests/submissions/Opera/responsetext-decoding.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: responseText decoding</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function request(type, input, output) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open("GET", "resources/status.php?content=" + input + "&type=" + type, false)
          client.send(null)
          //alert(encodeURI(client.responseText))
          assert_equals(client.responseText, output)
        }, document.title + " (" + type + " " + input + ")")
      }
      request("application/xml", "<?xml version='1.0' encoding='windows-1252'?><x>%FF<\/x>", "<?xml version='1.0' encoding='windows-1252'?><x>\u00FF<\/x>")
      request("text/html", "<!doctype html><meta charset=windows-1252>%FF", "<!doctype html><meta charset=windows-1252>\u00FF")
      request("text/plain;charset=windows-1252", "%FF", "\u00FF")
      request("text/plain", "%FF", "\uFFFD")
      request("text/plain", "%FE%FF", "")
      request("text/plain", "%FE%FF%FE%FF", "\uFEFF")
      request("text/plain", "%EF%BB%BF", "")
      request("text/plain", "%EF%BB%BF%EF%BB%BF", "\uFEFF")
      request("text/plain", "%C2", "\uFFFD")

      // XXX might need fixing
      request("text/xml", "%FE%FF", "")
      request("text/xml", "%FE%FF%FE%FF", "\uFEFF")
      request("text/xml", "%EF%BB%BF", "")
      request("text/xml", "%EF%BB%BF%EF%BB%BF", "\uFEFF")
    </script>
  </body>
</html>
